refactored db to seed with schema.sql and not with server reboot
