+++ package.json
@@ -9,7 +9,8 @@
   },
   "scripts": {
     "start": "node server/index.js",
-    "start:dev": "nodemon server/index.js"
+    "start:dev": "nodemon server/index.js",
+    "seed": "mysql -u root < schema.sql && node database/seed.js"
   },
   "dependencies": {
     "bluebird": "^3.5.3",
