+++ server/index.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const db = require('../database');
 const morgan = require('morgan');
+const db = require('../database/index');
 const url = require('url');
 
 const app = express();
@@ -9,16 +9,11 @@ const PORT = 3000;
 app.use(bodyParser.json());
 app.use(morgan('dev'));
 
-
-
-
 app.get('/book/:id', (req, res) => {
   const id = req.params.id;
   console.log(id);
   console.log('hello');
 });
-
-
 
 app.listen(PORT, () => {
   console.log(`listening on port ${PORT}`);
